+++ packages/stylelint-config-wordpress/scss.js
@@ -13,5 +13,6 @@ module.exports = {
     "at-rule-no-unknown": [ true, {
       ignoreAtRules: [ "extend", "at-root", "warn", "error", "if", "else", "for", "each", "while", "mixin", "include", "content", "return", "function" ],
     } ],
+    "scss/selector-no-redundant-nesting-selector": true,
   },
 }